ui(refactor): split 'superdesk-packaging' module.
