+++ scripts/superdesk-packaging/controllers/index.js
@@ -1,0 +1,5 @@
+export { PackagingController } from './PackagingController';
+export { SearchWidgetCtrl } from './SearchWidgetCtrl';
+export { CreatePackageCtrl } from './CreatePackageCtrl';
+export { PackageItemCtrl } from './PackageItemCtrl';
+export { CombinePackageCtrl } from './CombinePackageCtrl';
