+++ scripts/superdesk-packaging/directives/PackageEdit.js
@@ -1,0 +1,11 @@
+PackageEdit.$inject = ['authoring'];
+export function PackageEdit(authoring) {
+    return {
+        templateUrl: 'scripts/superdesk-packaging/views/sd-package-edit.html',
+        link: function(scope) {
+            scope.limits = authoring.limits;
+            scope._editable = scope.origItem._editable;
+            scope._isInPublishedStates = authoring.isPublished(scope.origItem);
+        }
+    };
+}
