+++ scripts/superdesk-packaging/services/index.js
@@ -1,0 +1,1 @@
+export { PackagesService } from './PackagesService';
