+++ scripts/superdesk-packaging/styles/packaging.less
