+++ scripts/superdesk-packaging/tests/packaging.spec.js
